Create Windows specific module
